+++ attributes/default.rb
@@ -1,2 +1,2 @@
-default['prometheus_exporters']['listen_interface'] = 'eth1'
+default['prometheus_exporters']['listen_interface'] = node['network']['default_interface']
 default['prometheus_exporters']['disable'] = false